requirements.txt
crawl4ai==0.5.0.post6
selenium==4.30.0
webdriver_manager==4.0.2
faiss-cpu==1.10.0
sentence_transformers==3.4.1